# Conflicts:
#   pkg/apis/cores/types.go
#   pkg/client-go/examples/create-update-delete/create-update-delete-Action/main.go
#   pkg/client-go/examples/create-update-delete/create-update-delete-Group/main.go
#   pkg/client-go/examples/create-update-delete/create-update-delete-Node/main.go
#   pkg/client-go/examples/create-update-delete/create-update-delete-Workflow/main.go
#   pkg/nodelet/task/monitor/group_monitor.go
#   pkg/nodelet/task/runtime/device/device.go
#   pkg/nodelet/task/runtime/runtime.go
#   pkg/nodelet/task/task_exporter.go

+++ pkg/nodelet/task/interaction/intwithWasm/grpc-client/wasm_client.go
@@ -2,11 +2,13 @@ package wasm_client
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"os"
 	"time"
 
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
 	"hit.edu/framework/pkg/component-base/logs"
 	wasm_interface "hit.edu/framework/pkg/nodelet/task/interaction/intwithWasm/proto"
 )
@@ -15,40 +17,59 @@ type WasmClient struct {
 	serverIPAndPort string
 	app             string
 	client          wasm_interface.WasmInterfaceClient
-	// conn            *grpc.ClientConn
-	ctx context.Context
+	conn            *grpc.ClientConn
+	ctx             context.Context
 }
 
-//TaskExporter直接与任务进行通信的client相关的函数接口，被assigner里的函数调用
-
-func NewClient(ctx context.Context, serverIPAndPort string) *WasmClient {
-	app := "wasm-test-demo"
-	return &WasmClient{serverIPAndPort: serverIPAndPort, app: app, ctx: ctx}
+func NewClient(ctx context.Context, port string, app string) *WasmClient {
+	// app := "wasm-test-demo"
+	return &WasmClient{serverIPAndPort: "127.0.0.1:" + port, app: app, ctx: ctx}
 }
 
-// 与任务建立连接
-func (c *WasmClient) Connect() error {
-	conn, err := grpc.Dial(c.serverIPAndPort, grpc.WithInsecure())
-	if err != nil {
-		logs.Error("wasm client:与任务建立连接失败")
-		return err
+// // 与任务建立连接
+// func (c *WasmClient) Connect() error {
+// 	conn, err := grpc.Dial(c.serverIPAndPort, grpc.WithInsecure())
+// 	if err != nil {
+// 		logs.Error("wasm client:与任务建立连接失败")
+// 		return err
+// 	}
+// 	grpcClient := wasm_interface.NewWasmInterfaceClient(conn)
+// 	logs.Info("wasm client created")
+// 	c.client = grpcClient
+// 	// c.conn = conn
+// 	return nil
+// }
+
+// 检查并重建与任务的rpc连接
+func (c *WasmClient) checkConnection() bool {
+	connState := true
+	if c.conn == nil {
+		// conn, err := grpc.NewClient(c.serverIPAndPort)
+		conn, err := grpc.Dial(c.serverIPAndPort, grpc.WithTransportCredentials(insecure.NewCredentials()))
+		if err != nil {
+			// logs.Debugf("failed to connect to Wasm grpc server:%v", err)
+			logs.Errorf("failed to connect to Wasm grpc server:%v", err)
+			connState = false
+		} else {
+			c.conn = conn
+			c.client = wasm_interface.NewWasmInterfaceClient(conn)
+		}
 	}
-	grpcClient := wasm_interface.NewWasmInterfaceClient(conn)
-	logs.Info("wasm client created")
-	c.client = grpcClient
-	// c.conn = conn
-	return nil
+	return connState
 }
 
 // 部署-grpc接口
 func (c *WasmClient) Deploy(wasm_file string) (*wasm_interface.Result, error) {
+	if !c.checkConnection() {
+		return &wasm_interface.Result{}, errors.New("Deploy: grpc connection failed")
+	}
 	// 目前的部署rpc接口是把wasm程序数据作为参数发送给运行时
 	// wasm code
 	// wasm_file_1 := "test/wasm/wasm_task/printf.wasm"
 	wasm_file_1 := wasm_file
 	f, err := os.ReadFile(wasm_file_1)
 	if err != nil {
-		logs.Error("read fail", err)
+		logs.Errorf("read fail %v", err)
 		return nil, err
 	}
 	// fmt.Println(f)
@@ -61,61 +82,85 @@ func (c *WasmClient) Deploy(wasm_file string) (*wasm_interface.Result, error) {
 	// deploy intent
 	deploy_intent := &wasm_interface.DeployIntent{Identity: "identity", Data: []*wasm_interface.Deploy{&deploy_data}}
 
-	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
+	// ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
+	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	result, err := c.client.Deploy(ctx, deploy_intent)
-	if err != nil {
-		logs.Error(c.app+" : deploy fail,", err)
+	for err != nil {
+		logs.Errorf("%v : Deploy() failed, retry to rpc Deploy() , %v", c.app, err)
+		result, err = c.client.Deploy(ctx, deploy_intent)
+		time.Sleep(500 * time.Millisecond)
 	}
-	out := fmt.Sprintf("deploy result: code: %d _ msg:  %s ", result.GetStateCode(), result.GetMsg())
+	out := fmt.Sprintf("wasm deploy result: code: %d _ msg:  %s ", result.GetStateCode(), result.GetMsg())
 	logs.Info(out)
+
 	return result, err
 }
 
 // init-grpc接口
 func (c *WasmClient) Init() (*wasm_interface.Result, error) {
+	if !c.checkConnection() {
+		return &wasm_interface.Result{}, errors.New("Init: grpc connection failed")
+	}
 	// init_intent_data
 	init_intent := &wasm_interface.InitIntent{App: c.app}
 
-	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
+	// ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
+	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	result, err := c.client.Init(ctx, init_intent)
-	if err != nil {
-		logs.Error(c.app+" : init fail", err)
+	for err != nil {
+		logs.Error("%v : Init() failed, retry to rpc Init() , %v", c.app, err)
+		result, err = c.client.Init(ctx, init_intent)
+		time.Sleep(500 * time.Millisecond)
 	}
-	out := fmt.Sprintf("init result: code: %d _ msg:  %s ", result.GetStateCode(), result.GetMsg())
+	out := fmt.Sprintf("wasm init result: code: %d _ msg:  %s ", result.GetStateCode(), result.GetMsg())
 	logs.Info(out)
 	return result, err
 }
 
 // start-grpc接口
 func (c *WasmClient) Start() (*wasm_interface.Result, error) {
+	if !c.checkConnection() {
+		return &wasm_interface.Result{}, errors.New("Start: grpc connection failed")
+	}
+
 	// init_intent_data
 	start_intent := &wasm_interface.StartIntent{App: c.app}
 
-	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
+	// ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
+	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	result, err := c.client.Start(ctx, start_intent)
-	if err != nil {
-		logs.Error(c.app+" : start fail", err)
+	for err != nil {
+		logs.Error("%v : Start() failed, retry to rpc Start() , %v", c.app, err)
+		result, err = c.client.Start(ctx, start_intent)
+		time.Sleep(500 * time.Millisecond)
 	}
-	out := fmt.Sprintf("start result: code: %d _ msg:  %s ", result.GetStateCode(), result.GetMsg())
+	out := fmt.Sprintf("wasm start result: code: %d _ msg:  %s ", result.GetStateCode(), result.GetMsg())
 	logs.Info(out)
 	return result, err
 }
 
 // destroy-grpc接口
 func (c *WasmClient) Destory() (*wasm_interface.Result, error) {
+	if !c.checkConnection() {
+		return &wasm_interface.Result{}, errors.New("Destory: grpc connection failed")
+	}
+
 	// init_intent_data
 	destroy_intent := &wasm_interface.DestroyIntent{App: c.app}
 
-	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
+	// ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
+	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	result, err := c.client.Destroy(ctx, destroy_intent)
-	if err != nil {
-		logs.Error(c.app+" : destroy fail", err)
+	for err != nil {
+		logs.Error("%v : Destory() failed, retry to rpc Destory() , %v", c.app, err)
+		result, err = c.client.Destroy(ctx, destroy_intent)
+		time.Sleep(500 * time.Millisecond)
 	}
-	out := fmt.Sprintf("destroy result: code: %d _ msg:  %s ", result.GetStateCode(), result.GetMsg())
+	out := fmt.Sprintf("wasm destroy result: code: %d _ msg:  %s ", result.GetStateCode(), result.GetMsg())
 	logs.Info(out)
 	return result, err
 }
